modified statistical functions to adapt long long int.

+++ time_changed_barrier_option2/include/statistical_functions.h
@@ -6,6 +6,6 @@ void MeanVarianceNext(double *running_mean_ptr, double *running_variance_ptr, un
 
 enum AcceptanceLevel {NinetyFive, OneOverAHundred, OneOverAThousand};
 
-unsigned long SampleSizeGenerator(double accuracy, double variance, statistical_functions::AcceptanceLevel level);
+unsigned long long SampleSizeGenerator(double accuracy, double variance, statistical_functions::AcceptanceLevel level);
 
 }//namespace statistical_functions
